feat: add tests for channel.
