+++ casbin/casbin.vcxproj
@@ -153,6 +153,7 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="channel.cpp" />
     <ClCompile Include="config\config.cpp" />
     <ClCompile Include="duktape\duktape.cpp" />
     <ClCompile Include="effect\default_effector.cpp" />
@@ -204,7 +205,6 @@
     <ClCompile Include="util\array_remove_duplicates.cpp" />
     <ClCompile Include="util\array_to_string.cpp" />
     <ClCompile Include="util\built_in_functions.cpp" />
-    <ClCompile Include="util\channel.cpp" />
     <ClCompile Include="util\ends_with.cpp" />
     <ClCompile Include="util\escape_assertion.cpp" />
     <ClCompile Include="util\find_all_occurences.cpp" />
@@ -217,6 +217,7 @@
     <ClCompile Include="util\trim.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="channel.h" />
     <ClInclude Include="config.h" />
     <ClInclude Include="config\config.h" />
     <ClInclude Include="config\config_interface.h" />
@@ -286,7 +287,6 @@
     <ClInclude Include="rbac\role_manager.h" />
     <ClInclude Include="util.h" />
     <ClInclude Include="util\built_in_functions.h" />
-    <ClInclude Include="util\channel.h" />
     <ClInclude Include="util\pch.h" />
     <ClInclude Include="util\util.h" />
   </ItemGroup>
