+++ casbin/util.h
@@ -19,6 +19,5 @@
 
 #include "./util/built_in_functions.h"
 #include "./util/util.h"
-#include "./util/channel.h"
 
 #endif
