+++ test/test.vcxproj.filters
@@ -51,6 +51,9 @@
     <ClCompile Include="test_enforcer_cached.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="test_util_channel.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pch.h">
